Move `LegacyDeviceFile` tests in its own module
