+++ oxidation/libparsec/crates/client_types/src/lib.rs
@@ -5,6 +5,8 @@ mod local_device;
 mod local_device_file;
 mod local_manifest;
 mod path;
+#[cfg(test)]
+mod test;
 
 pub use error::*;
 pub use local_device::*;
